update adapter tests
